Move csv dialog to utilities
